[Chore #161695595] Relocate configurations to proper file
